requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
tensorflow-cpu>=2.16.0
pillow>=10.0.1
numpy>=1.26.4
pydantic>=2.4.2
google-genai
gTTS>=2.4.0
python-dotenv>=1.0.0
opencv-python-headless>=4.10.0.84
